Extract and refactor cell renderer
